requirements.txt
# 小红书热门博客分析系统依赖包

# Web框架
Flask==2.3.3
Flask-CORS==4.0.0

# 数据处理
pandas==2.1.4
numpy==1.24.3

# 网络请求
requests==2.31.0
urllib3==2.1.0

# 网页解析
beautifulsoup4==4.12.2
lxml==4.9.3

# 浏览器自动化
selenium==4.15.2
undetected-chromedriver==3.5.4
webdriver-manager==4.0.1

# 用户代理
fake-useragent==1.4.0

# 环境变量
python-dotenv==1.0.0

# 图表和可视化
matplotlib==3.8.2
seaborn==0.13.0

# 日期时间处理
python-dateutil==2.8.2

# 正则表达式增强
regex==2023.10.3

# 进度条
tqdm==4.66.1

# 日志处理
loguru==0.7.2

# 配置文件处理
PyYAML==6.0.1

# 数据验证
pydantic==2.5.0

# 异步支持
aiohttp==3.9.1
asyncio==3.4.3

# 文件处理
pathlib2==2.3.7

# 加密和安全
cryptography==41.0.8

# 图像处理（可选）
Pillow==10.1.0

# 数据库支持（可选）
sqlite3

# 开发工具（可选）
black==23.11.0
flake8==6.1.0
pytest==7.4.3 